Stop intuiting package names
This originally came in as a part of 99cc6a7 as rpt was being
refactored to have fewer fatal conditions and to produce more
normal output. It turns out that it's VERY useful to npm to be
able to distinguish between "no package.json" and "package.json
with only name=dirname".