Refactor to be promise-based
This commit changes no functionality, except that a Promise is returned
by the read-package-tree function.  (A supplied callback is attached to
the returned promise if provided.)

Test coverage is brought up to 100%, and some code paths have been
slightly optimized, but the excessive level of realpath/lstat calls
remains for now.