Revert breaking change, preserve legacy design bug
Two breaking changes were mistakenly landed in v5.3.0.

The first is returning full paths on the Node.path property, where
relative paths were previously provided.  (Since the fully resolved path
is on the Node.realpath property, it's not really necessary to have the
full path on Node.path anyway, and this was just an oversight.)

The second is that the children of a Link node were previously rooted
with the *Link* as the parent node, rather than the target.  This causes
problems in dependency tree analysis, which npm CLI mostly (but not
entirely) works around.  When combing through this code to remove the
fs.realpath race conditions, this weird bug was 'fixed', but doing so
breaks a lot of the code that works around this read-package-tree design
flaw (which dates back to before this code was abstracted out of npm CLI
4 years ago, so working around it as established context can certainly
be forgiven).

Version 6 will fix this long-standing design bug, but doing so requires
much more work on the tree shaking algorithms in npm CLI.

Fix https://github.com/npm/read-package-tree/issues/21