Clean up path and parent handling
Make node.parent a setter that updates location, so that cases where a
package is linked multiple times throughout the tree resolves to the
appropriate path.  If a link target is outside the tree, then node.path
remains null, rather than being set to the logical path where it first
occurs in the tree.

Fix #1