New classes for Node, Edge, and Link
This is taking the module completely away from the approach used by
read-package-tree.  Loading is still not done, so the tests are
expectedly failing, but the need for this rewrite became more and more
apparent when trying to shoehorn virtual loading and tree munging into
the older data structure.

With this, the dependency graph is automatically generated and updated
as the folder tree is defined or mutated, making it much easier to do
the kinds of operations that will be required for npm's use cases.

From the Node point of view, there's only one kind of 'thing' that
represents a dependency, which is captured in the Edge class, making it
much simpler to keep up to date when re-parenting nodes.