Calculate dep flags (dev, optional, etc.) properly
This also adds the capability of a Node constructor to not be given aa
`path` or `realpath`, since that can be inferred when assigning to a
parent.  In that case, it gets the `name` field from its `pkg` option.

Of course, this means that alias deps will have to be given a path (at
least `path: 'alias-name'`) when being created.  Another option might be
to have an explicit `name` field in the options.