Resolve deps to fs parent folder children
Fix: #15

Note: this does NOT support pnpm style trees.  They still show up as a
ton of missing deps.  But, this commit adds a test fixture and snapshot
for how pnpm trees are handled, so we can track changes against it more
easily.

pnpm trees will be addressed via #16

This lays the groundwork for Workspaces, by resolving dependencies in
the `packages` (or whatever) subfolders against the root node, if those
deps are found there.