Replace Metadata class with Shrinkwrap class
This unites the handling of npm-shrinkwrap.json and package-lock.json
files into a single class, and defines the lockfileVersion 2 style
shrinkwrap files.

There's no need for an .arborist-lock.json file in node_modules, because
the 'packages' section covers all that we would put in there anyway.

In lockfileVersion 3, we can drop the dependencies section, along with
the legacy metadata code supporting it.  (Likely we'll want to drop the
writing/tracking of the legacy metadata before we drop support for
reading it, if ever.)

The loadVirtual function has been removed, but will be brought back
similar to how loadActual was migrated into the Arborist class.  The
intent of the Shrinkwrap class is to make that easier to accomplish,
since we effectively just have to generate all the Node and Link objects
and assign their parents to hook them up into a tree, and the lockfile
will contain the arguments needed to generate them.

Still todo: store funding info in the lockfile, so that we can run `npm
fund` without loading the entire actual tree.  There are a few other
other use cases where we'd want to get a quick and dirty snapshot (ci
and audit, at least).