YarnLock: class to parse and stringify yarn lock file metadata
This is just the initial implementation spike.  It does not yet provide a nice
way to lookup or update metadata, or keep it up to date with a Node tree.

Next step: integrate with the Shrinkwrap class so that a yarn.lock file is
automatically respected and kept in sync if it exists.