Get metadata from yarn.lock, clean up root/parent behavior
This fixes a lot of weird edge cases and one race condition in how
metadata was being handled between the root, top, and parent semantics.

The race condition was that loadFSTree would sometimes be called on a
link whose target had not yet loaded, resulting in an empty children
collection.