DRY path/location handling, fix deeply nested edge case
DRY the code that was handling metadata fetching and updating, and
take care of an edge case that prevented deeply nested link targets
from being properly updated when they eventually get parented.

This is mostly a refactor that just cleans up the code for readability,
but it does fix a (albeit obscure and unlikely) bug in the process.