Resolve confusion around 'root' vs 'path'
On Node and Link objects 'root' refers to the root node.

But on Arborist and Shrinkwrap objects, 'root' referred to the root
path.

This led to some weird and confusing internal terminology.  Fix by
making 'path' always refer to the string path to a thing, and reserving
'root' to refer to a root node in a graph.