Shrinkwrap class takes an options object
The number of things that had to be passed to the Shrinkwrap class was
getting a little out of hand.  We're contemplating adding a 'backup
lockfile' concept, which will make matters even worse, and it needs to
know about save:false, so the right thing to do is have it take a set of
options at construction time, rather than passing booleans around.