keep path resolution relative to file storing it
This changes how lockfiles store the resolved paths for Link nodes.

Instead of making it relative to the Link, the path is always relative
to the lockfile in question.  That gets rid of an extra layer of
indirection, and makes it easier to move Link nodes around in the tree
without having to do lots of path calculations that can be hard to
reason about.

Note: this makes some other tests break, but they will be updated in a
subsequent commit.