Add node replacement, fix node.resolved for file:
This adds the replace(), replaceWith(), canReplace(), and
canReplaceWith() methods to the Node class, and some other methods
needed for building ideal trees.

It also adds consistency to how shrinkwraps and node objects track the
node.resolved values for file: deps.  The resolved value is now always
relative to the object or file reporting it.  That means that we
effectively can't do anything with the _resolved value in package.json
for file: deps, so we have to rely on package-lock and shrinkwraps for
those.