fsParent and target are bidirectional and managed
This automatically updates the path, realpath, and associated fields
whenever the node.fsParent or link.target are modified, as we were
previously doing with node.parent.

That is, if the link.target is reassigned, then link.realpath is updated
to match link.target.path.  If node.fsParent is modified, then its path
is updated to be relative to the new fsParent's path.  And, if changed,
the paths of all nodes referencing this one as either a parent or
fsParent or link target are similarly updated.

In order to accomplish this, the linksIn and fsChildren sets are added,
making link.target and node.fsParent bidirectional.