Properly dedupe nested meta-dep cycles
This implements a pruning algorithm when replacing nodes, such that any
cyclical set of dependencies required by the old version can be
optimally replaced by the dependencies of the new version.

For example, a cycle of peer deps, all updated to a new version with the
new version of the dep being updated, would cause conflicts when finding
the ideal place in the tree, because each node would have edgesIn from
another conflicting version.  As a result, the cycle of deps would get
nested, and then after placing everything, the pruning step would remove
those now-extraneous deps anyway.

The earlier pruning step in the case of node replacement reduces the
cases where we have extraneous deps left over in the first place (maybe
eliminates it altogether? not sure, haven't taken the time to prove it
one way or another), so that these conflicting metadeps will be moved
out of the way before attempting to place dependencies.

Additionally, a check is now performed after placing a node to find any
nodes by the same name which are descendants of the target and can be
safely replaced by the target.  So, if a previous installation required
a nested placement of a dep, but it's no longer required after an
update, then the nested node is duduped out automatically as part of the
subsequent placement.