diff: do not clobber fsChildren
TODO: provide a deep:true argument to tell it to diff the children of
fsChildren nodes.

There's no sense in diffing these, however.  Either they're included in
a package's contents (in which case, they'll be handled when/if the node
is replaced or otherwise modified), or they're a separate project and
shouldn't be modified.