Remove race in reify test
npm v6 only inconsistently saves the resolved url for deps that
are bundled.

That's fine, since we don't have to fetch them anyway, but it does
mean that the test had a race, where it would get a different
snapshot based on whether it had loaded that node (and thus updated
the shrinkwrap) or not.