Split Arborist class into 4 mixins
This change is mostly for esthetic code management, but also it makes
the following substantive changes visible in the API:

- Internal methods and properties moved to hidden Symbol properties.
  (Some are shared symbols for testing or sharing between mixin
  classes.)  This prevents accidental clobbering, and makes it much
  clearer what is intended for external consumption.
- legacyBundling and preferDedupe were moved to buildIdealTree/reify
  options instead of constructor options.

The test coverage map was simplified, so that now the various Arborist
method tests only test the parts that they're intended to.  (There were
some buildIdealTree and loadVirtual code paths that were only getting
hit by the reify test, but now it's all properly covered by what it's
intended to be covered by.)

Tests were also added for the cases where rimraf fails.  Even though
this is unusual, it's not impossible.