Ensure that moving a child node refreshes all edges
Previously, it was only refreshing *invalid* edges, meaning that any
previously-valid edges would be referencing nodes outside the tree when
the parent and root were changed.  This led to some weird duplication in
the reified actualTree, where nodes were left with edgesIn coming from
the temporary actualTree of bundleDependencies nodes.