Set resolved property properly on Link nodes
Should be the relative path from the link's containing folder, not from
the link path itself.  Ie, it should be the actual argument passed to
fs.symlink() if one were to try to reify the Link node.

This changes a lot of snapshots, but the only one that's really relevant
is the handling in the YarnLock class, since all the other instances of
using consistentResolve to re-relativize a path were handling Link nodes
specially anyway (eg, in reify, shrinkwrap, etc.)