Filter fsevents out of snapshots so Travis doesn't fail
The reify() method now returns the *actual* actualTree, so even though
engine-mismatching optional deps remain in the package-lock, they're not
being passed in the resolved tree.

Filter the offending darwin-only module out of the snapshots so that Travis
(which is Linux) doesn't fail tests.  In that case, just manually verify that
fsevents is actually in the package-lock, even though it isn't in the snapshot.