further simplify adding new deps to the root
The previous revision of the buildIdealTree/reify `add` option converted
the package.json-esque data structure into instead passing a list of
specs for each location in the package, since the name couldn't be known
100% of the time up front.

However, that still puts too much of the burden of placing deps onto the
caller.

When you run `npm install foo`, if `foo` is already a devDependency,
then the updated save spec gets added there, unless specified otherwise.
Since we aren't guaranteed to know the _name_ of the dep until we've
resolved it, how is the caller (ie, the npm CLI) supposed to know where
to put it in the objecet?

At the expense of a slight amount of flexibility, this changes things
further so that the `add` option mirrors exactly the positional
arguments passed to the `npm install` command: it's just a list of
specs, that's it.

Two more options were added:

- saveType
- saveBundle

If the saveType is not set, then added deps are added wherever they
already exist, or to `dependencies` if they're not already present.
Otherwise, it should be set to one of the edge type values:

- 'prod'
- 'dev'
- 'optional'
- 'peer'
- 'peerOptional'

If `saveBundle` is true, then newly added deps are added to the
`bundleDependencies` list as well.