Correct logic to leave bundled deps untouched
This was a really interesting bug to track down.

Basically the bug comes down to a way that arborist tracks bundled deps.
The getter `Node.isBundled` calls `Node.getBundler()`, which tries to
figure out which node is bundling this one, if any.  This uses the
following algorithm:

- if there's no parent node, it's definitely not bundled
- if the parent node is bundled, then the bundler is the parent's
  bundler.  (if you bundle `a`, you also bundle `a/node_modules/b`)
- if the parent package has a `bundleDependencies` list, and the node's
  name is on the list, then it's bundled, and the parent is the bundler
- if any of the edgesIn referencing this node are bundled, and _their_
  bundler is the node's parent, then that is the bundler of this node as
  well.

The downside of this algorithm is that the 4th step of checking the
edges can potentially be slow (like, _O(n!)_ slow) but since it can only
ever end up traversing the edges within a given set of deps at a single
level of the tree, n is always very small.  even absurd bundled dep
lists still only have a few dozen to a hundred or so bundled deps, and
it'd only get pathological if you had one bundled dep with a long chain
of dependencies hanging off of it, all of which also depended on one
another (but were not explicitly bundled).

The benefit is that if we remove the thing that is triggering the
bundling, everything else that it was pulling in automatically stops
being marked as a bundled dep, without having to re-evaluate _bundling_
every time we mutate the tree (at the cost of having to evaluate the
bundling triggered by edges every time we check if one of these
meta-deps is bundled, but as established, this is not very expensive).

That's all fine and good, _until_...

There's a shortcut in the diff calculation that basically says, if a
node is unchanged, and it has bundled children, then we want to mark all
the children as unchanging as well.  A Diff object has `actual` and
`ideal` node references, but the `ideal` wont' know about the unlisted
metadeps in the bundle until it actually unpacks.  So, the hacky kludge
is, let's just move all the bundled `actual` nodes over to be
re-parented on the ideal node.  This would modify the `actual` tree, of
course, but since we don't bother to actually unpack anything that's a
bundled child of an unchanging node, it doesn't matter.

However, I was doing it like this:

```js
for (const [name, node] of actual.children.entries()) {
  if (node.inBundle)
    node.parent = ideal
}
```

Which means, if a metadep is bundled, but comes _after_ the primarily
bundled dep in the `children.entries()` list, then first we'll move over
the direct dep, _and then_ hit the metadep.  At which point, it says
"nope, not in the bundle".

Since the ideal tree doesn't contain that metadep (since it didn't
resolve the bundled deps), and will never resolve that metadep (since it
doesn't traverse bundled children, as their deps should already be in
the bundle), the resulting diff will show that there is a node in the
actual tree that _isn't_ found in the ideal tree, _and delete the
metadep!_

Solved by making the hack loop work like this:

```js
const bundledChildren = []
for (const [name, node] of actual.children.entries()) {
  if (node.inBundle)
    node.parent = ideal
}
for (const node of bundledChildren) {
  node.parent = ideal
}
```

Turns out I already had a test for this, and once I realized wtf was
going on, lo and behold, the `tap-snapshots/test-diff.js-TAP.test.js`
file shows that it's dutifully deleting the
`bundler/node_modules/metabundled` node.  Now that it's fixed, it'll
show up as a snapshot error if we break it again.