test: asymmetrical bin across versions repro
Added a test that reproduces a problem when installing transitive deps
that contains multiple versions in which its `bin` package.json
declaration is not symmetrical across deduped versions.

When a lockfile v1 is available the installation will fail to produce
`node_modules/.bin` symlinks. A similar install setup but with an absent
lockfile will yield a functioning bin folder.