Arborist.audit() with fix support
Still lacking proper unit tests for buildIdealTree.  That is only
exercised at the moment by Arborist.audit({fix: true}), through the
reify call.

It would be good to stash the ranges to be avoided somewhere, either in
the lockfile or in a separate file, so that we can support the case
where someone runs `npm audit` followed immediately by `npm install`,
without giving them things that we ought to know are no good.