Tests for buildIdealTree's audit behavior
This also adds the functionality of having loadVirtual attempt to read
the package.json file for link targets found in legacy shrinkwraps,
because otherwise their dependencies were never being loaded.  This
means that the recheckEngineAndPlatform in Arborist.reify() is no longer
necessary for link targets, but we still have to do that for
non-link-target nodes, since it also does not set bin/engines/os/cpu
fields.