Produce JSON-friendly format for AdvisoryReport
This adds a toJSON method on the AdvisoryReport class, so that it can be
passed to npm-audit-report, without that module having to know about
Arborist's internals.

This also moves the 'fixAvailable' calculation into the AdvisoryReport
class rather than buildIdealTree, since it's relevant to the report that
we output for the user, and thus needs to be in the pojo result.

Also, remove some tap snapshot files for tests that were removed (in
some cases, quite a while ago!)