Fix case where dep vuln also has an advisory
This preserves the information that was previously being thrown aside,
where a node that is vulnerable by virtue of having a dependency on a
vulnerable module *also* has an advisory posted against it.

Prior to this commit, the dep vuln would be lost.

Also, more importantly, the calculation for upgrading severity was
completely backwards, which could result in vulnerabilities being
reported at a lower level then they should have been in these cases.