Do not use --before option when looking for audit fixes
Otherwise, we end up getting ETARGET errors, because the version
depended on likely does not exist with the time specified.

Also, whatever errors we DO get from pickManifest when looking for
vulnerability fixes, we should just treat as a reason to avoid that
version, since it may well not be installable anyway.