tracker: be more lenient
There are cases when doing an 'audit fix' with a lot of outdated deps,
where a node might be scheduled for processing, but then be removed
prior to getting to it in the queue, without putting another node in
that node.location.  As a result, in normal operations, the
node.location is not as permanent or deterministic as we might hope,
resulting in a lot of spurious throws for unfinished trackers and
nonexistent subtrackers.

We really need to re-evaluate how this is being done.  It's not the best
user experience, since the ideal tree building process is not
predictable or linear enough to lend itself to a progressbar display.
In the meantime, at least we can stop it from throwing.