include pkg name/version in loadVirtual test snapshots
Demonstrates root cause problem of #70.  The Shrinkwrap class only adds the
'name' field for new lockfiles, but loadVirtual expects it to be there

Really, it probably shouldn't be there for v2 lockfiles either, since it's
almost always extraneous, except for alias dep types, but regardless,
loadVirtual should work out the package name from the folder if it's not
present.