Base idealTree on actualTree if no lock present
Fix #79

If there is no package-lock.json or npm-shrinkwrap.json file present,
then base the idealTree on the actualTree if available.

This minimizes surprising meta-dep updates (and in some cases, extremely
slow install times!) when not using a package-lock.json file.

A caveat here is that it _may_ result in a lockfile being saved for the
first time with `resolved: null` for some deps, if they were present in
the `node_modules` folder to start with.