add missing package.json in test fixture
This was causing a weird as hell race condition due to the prior commit
that checks if the hidden lockfile is up to date.

If the test fixture was created with a >10ms gap between the hidden
lockfile being created and the package contents being created, then the
hidden lockfile wouldn't be respected, and it'd have to load the
package.json from disk.

However, since the package.json was missing, this meant that the semver
node always had a node.package of {}.

That timing issue would only hit if the test fixture generation took
more than 10ms, AND the delay landed in just the right spot in the test.
Reproduced by adding a ton of loging, and then running the test in a
tight loop along with 6 other tests going at the same time, and even
then, it only usually triggered after 5-10 minutes.

It wasn't a problem before, because a hidden lockfile contains
everything necessary to load the tree.  So, when there was no safety
check to ensure that the hidden lock is newer than the packages it
describes, the lack of a package.json file was never a problem.