test: ensure hidden lockfiles in fixture are newer than contents
There's a race in the fixture generation we use which can _sometimes_
end up creating a hidden lockfile isn't newer than the contents of the
node_modules folder, and ends up being ignored out of an abundance of
safety.

Ensure that doesn't happen in our tests by bumping the mtime after
fixture creation.

Also: add some missing fixture folders that were present in the
reify-cases snapshots, but not in the actual directories.