node: reload edgesIn to nodes matching name in reloadNamedEdges
Wow, that's a dense commit heading.  Tried hard to get it under 50
chars, and still failed.

When we add a node into a parent node, we call reloadNamedEdges on the
parent, to walk the tree reloading any edges that have the name of the
child node being added.  This takes care of the case where a node is
being added, and will end up changing the dependency resolution of nodes
within that portion of the tree.

We were previously skipping over any nodes by that same name, and their
children, with the reasoning that they would block the new higher-level
addition from being reached by their children anyway.

However, this turns out to not be sufficient, because those nodes may
still have edgesIn that have to be re-validated, and would not otherwise
be reached by the walker, if there is one version of a given named node
nested inside a _different_ node with the same name.

This becomes relevant for detecting infinite regressions in the nesting
algorithm, because we need those invalid links to show that there's
still more work to do, or else we never hit the infinite regress, and
cannot warn the user that their tree is impossible to resolve.