Do not incorrectly report bundleDependencies as fixable
Rather than attempt to fix bundle dependencies in-place, report the
dependent module as vulnerable if its dependency range _intersects_ with
the vulnerable range, since it _may_ contain that version in its bundle.

From there, we can simply report the vulnerability as we normally do,
with the assurance that fixing it will certainly avoid the vulnerable
dependency.

Note that we don't actually have to report them as unfixable (as long as
some non-bundled and/or non-intersecting dependency range is found), or
track the bundled-ness through to the data passed to
npm-advisory-report.

This also fixes an error where bundleDependencies in `loadActual` were
too liberally set.  Only the top-most bundled parent should be
considered a bundleDependency of its parent, not every child node.  This
showed up in being overly broad in determining which dependency versions
are affected by a bundled meta-dep.

Fix #69