audit: use new bulk advisory endpoint
This isn't available on the public registry yet, but tested against
live data on the internal npm staging registry.  When it goes live,
it'll make audits significantly lighter.

When we _do_ fetch data from the quick audit endpoint, because the
bulk advisory endpoint is not available, strip out all the info that
we don't use for anything.

Fix #101