Do not use hidden lockfile if a package is missing
If a hidden lockfile contains a package which is missing from the
node_modules folder, then it should be considered invalid.

Typically, this would already be caught by the hidden lockfile
being created prior to the folder being created.  However, in the
bundle deps bug fixed in the previous patch version, the lockfile
at node_modules/.package-lock.json would be created afterwards,
and packages in node_modules would simply not be created in the
first place (so there's no time update for deleting an entry).