Do not remove bundled deps when updating pkg
When calculating the diff between the actual and ideal, the ideal tree
does not contain bundled dependencies, since they'll be installed
automatically and thus don't have to be reified.  As a result, the Diff
shows these nodes as "removed", since they exist in the actual tree, but
not in the ideal tree.

However, this causes a problem when updating a package!

When reifying, we add all the "removed" nodes to the trash list, and
remove them at the end of the process.  In this case, it's unnecessary,
since the contents of the updated package will have already been moved
aside, including all their bundled deps.  The items on the trashlist are
thus the _new_ installed bundled deps of the new version of the package.

This led to a particularly perplexing situation self-installing the npm
cli, where it would appear to install properly, but installing again
left it without any dependencies at all!

The fix is relatively straightforward, but it does mean that the Diff
for a normal reify will be somewhat misleading, showing many removals
that are not actually going to be removed.  The diff from a
buildIdealTree({complete:true}) tree will be accurate, but at the
expense of extra unpacks for all nodes with bundleDependencies.